Refactoring
* Make Patches and Queries have two generic parameters to
express the original type as well as the resulting type.

* Create a CompoundQuery class to allow for any arbitrary
queries to be combined into a single query.

* Remove the second generic type in BaseStore. For the most
part, the type of the items in the store will not change
between the source and the subStore, so they can use the
same generic type.

If they did change types, in one case the
items could be mapped to a completely different type. In
this case updates can no longer happen by delegating to
the parent because there's no way to reverse engineer
the relationship between the items.

In the case where a selection or a similar operation has
been performed, it would be useful, but the direction of
the type relationship is contravariant. If the source has
Items of type {a: string, b: string}, the child would have
elements of either {a: string}, {a: string, b: string}, or
{b: string}. In any other case we have the problem mentioned
above. But in this case, the resulting item is a superclass
of the source store. Which means the signature for this
opertion would need to be something like
```
Store<A, B extends A> {
	source: Store<B>
	select<C extendedBy A>: Store<C, A>;
}
```
But we can't express that.