Minor refactor and more tests
* Minor changes to update interface so that all updates can
optionally have an item or index property. This avoids a lot
of casting to different update types.

* Updates to track and release API so that they return a
promise to the store.

* Don't send out an initial update of current items when
observing the store. If using fetch, it's not performant,
and if not using fetch, the results can be surprising
because the local data may not be what is expected at
the time of observation.

* Getting rid of fat arrow functions to match dojo style
guide

* Making the order of actions more controllable via custom
store managers by adding a retry function. The default
manager currently just executes the next available action
asynchronously as soon as the current action executes once.
Any retries on that action are treated as new actions and
just queued normally. With the separate retry method an
action manager could be implemented that waits until the
current action is completed, including waiting for any
retries.